# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Imposter AI API
  version: 1.0.0
  description: |
    # Imposter AI — Backend Specification

    Backend API для игры Imposter AI — управление категориями, AI-генерацией слов, premium-контентом и remote config.

    ---

    ## Сущности

    ### Category
    Описывает тематическую категорию слов. Делятся на free и premium через флаг `is_premium`.
    Активность управляется через `is_active` — скрывает без удаления. `sort_order` определяет порядок в клиенте.
    Все текстовые поля хранятся как LocalizedString — JSON-объект с ключами по ISO locale.

    ### WordPack
    Набор слов привязанный к категории. Создаётся отдельно под каждую локаль — автоматического перевода нет.
    - Для стандартного режима `impostor_word = null` — impostor не получает слово.
    - Для Party Mode (Undercover) `impostor_word` содержит похожее но отличное слово.
    - При запросе клиент передаёт `locale`, сервер ищет подходящий пак; если нет — fallback на `en`.

    ### FeatureConfig
    Управление feature flags и контент-гейтингом. Клиент получает весь конфиг одним запросом `GET /config` при старте.
    Флаги кешируются на клиенте с TTL 15–30 мин. `payload` может содержать лимиты AI-запросов для free-tier.
    Ключи: `"ai_theme"`, `"premium_categories"`, `"party_mode_extra_roles"`.

    ### AITopicRequestLog
    Лог запросов на AI-генерацию — для дебага, модерации, rate limiting. Хранится для аудита, не возвращается клиенту.
...

    ### LocalizationContent
...

    ---

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Анонимный идентификатор устройства, генерируется при первом запуске
    AppVersion:
      name: X-App-Version
      in: header
      required: true
      schema:
        type: string
        example: "1.0.0"
    LocaleHeader:
      name: X-Locale
      in: header
      required: false
      schema:
        type: string
        example: "ru"
      description: ISO 639-1 код языка
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: RATE_LIMIT_EXCEEDED
            message:
              type: string
              example: AI request limit reached for today
            details:
              nullable: true
              example: null
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: animals
        name:
          type: string
          example: Animals
        description:
          type: string
          example: Guess the animal
        is_premium:
          type: boolean
        is_active:
          type: boolean
        cover_image_url:
          type: string
          nullable: true
          example: https://cdn.../animals.png
        sort_order:
          type: integer
          example: 1
    CategoryPremium:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            preview_words:
              type: array
              items:
                type: string
              example: ["Matrix", "Avatar"]
    WordEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        civilian_word:
          type: string
          example: Elephant
        impostor_word:
          type: string
          nullable: true
          example: null
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
          example: []
...
paths:
  /categories:
    get:
      summary: Список всех активных категорий
      description: Возвращает все активные категории (free + premium) для отображения в пикере.
      tags: [Categories]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: en
          description: ISO locale. Default en
      responses:
        '200':
          description: Список категорий
          headers:
            Cache-Control:
              schema:
                type: string
                example: max-age=1800
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
              example:
                categories:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    slug: animals
                    name: Animals
                    description: Guess the animal
                    is_premium: false
                    cover_image_url: "https://cdn.../animals.png"
                    sort_order: 1
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    slug: cinema
                    name: Cinema
                    description: Movies & TV
                    is_premium: true
                    cover_image_url: "https://cdn.../cinema.png"
                    sort_order: 10
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories/premium:
    get:
      summary: Список premium категорий с preview
      description: Возвращает только premium категории с preview_words. Используется на paywall-экране.
      tags: [Categories]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Список premium категорий
          headers:
            Cache-Control:
              schema:
                type: string
                example: max-age=1800
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryPremium'
              example:
                categories:
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    slug: cinema
                    name: Cinema
                    description: Movies & TV
                    is_premium: true
                    cover_image_url: "https://cdn.../cinema.png"
                    sort_order: 10
                    preview_words: ["Matrix", "Avatar"]
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories/{category_id}/words:
    get:
      summary: Слова для категории
      description: |
        Возвращает слова для конкретной категории. Для premium-категорий проверяет is_premium в профиле пользователя.
        Если WordPack для запрошенной локали не найден — сервер возвращает fallback на `en`, не ошибку.
      tags: [Categories]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID категории
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: en
          description: ISO locale. Fallback en
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [standard, party]
            default: standard
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1
            maximum: 5
          description: Количество слов для раунда
      responses:
        '200':
          description: Слова категории
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              example: 100
            X-RateLimit-Remaining:
              schema:
                type: integer
              example: 99
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp когда освободится первый слот
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: string
                    format: uuid
                  locale:
                    type: string
                  mode:
                    type: string
                    enum: [standard, party]
                  words:
                    type: array
                    items:
                      $ref: '#/components/schemas/WordEntry'
              example:
                category_id: "550e8400-e29b-41d4-a716-446655440000"
                locale: en
                mode: standard
                words:
                  - id: "550e8400-e29b-41d4-a716-446655440010"
                    civilian_word: Elephant
                    impostor_word: null
                    difficulty: easy
        '403':
          description: Пользователь не имеет premium-подписки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Категория не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /config:
    get:
      summary: Remote config приложения
      description: |
        Возвращает полный remote config приложения.
        Запрашивается при каждом cold start. Кешируется локально с TTL 15–30 мин.
        При недоступности сервера клиент использует last known good config.
      tags: [Config]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
      responses:
        '200':
          description: Remote config
          headers:
            Cache-Control:
              schema:
                type: string
                example: max-age=900
          content:
            application/json:
              schema:
                type: object
                properties:
                  features:
                    type: object
                  content:
                    type: object
                  app:
                    type: object
                  updated_at:
                    type: string
                    format: date-time
              example:
                features:
                  ai_theme:
                    enabled: true
                    free_daily_limit: 5
                    premium_daily_limit: 50
                  premium_categories:
                    enabled: true
                  party_mode_extra_roles:
                    enabled: true
                    is_premium: false
                  import_custom_words:
                    enabled: true
                content:
                  default_free_categories: ["animals", "food", "movies"]
                  featured_category_slug: cinema
                app:
                  min_supported_version: "1.0.0"
                  force_update_version: null
                  maintenance_mode: false
                updated_at: "2026-06-06T12:00:00Z"
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /localizations:
    get:
      summary: Серверные строки локализации
      description: |
        Строки локализации для контента требующего обновления без релиза приложения
        (paywall тексты, промо, onboarding). Включён в scope MVP.
      tags: [Config]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
        - name: locale
          in: query
          required: true
          schema:
            type: string
            example: ru
          description: ISO locale
        - name: keys
          in: query
          required: false
          schema:
            type: string
            example: "paywall.title,paywall.subtitle"
          description: Comma-separated список нужных ключей. Если пусто — все
      responses:
        '200':
          description: Строки локализации
          headers:
            Cache-Control:
              schema:
                type: string
                example: max-age=3600
          content:
            application/json:
              schema:
                type: object
                properties:
                  locale:
                    type: string
                  strings:
                    type: object
                    additionalProperties:
                      type: string
              example:
                locale: ru
                strings:
                  paywall.title: "Открой всё"
                  paywall.subtitle: "AI-темы, премиум-категории и многое другое"
                  onboarding.step1.title: "Раздаём роли"
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...